problem: apis cannot be customized because have a single init point solution: split connection config and actual apis into separate classes
