+++ src/main/java/io/emeraldpay/api/Chain.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.emeraldpay.grpc;
+package io.emeraldpay.api;
 
 public enum Chain {
