+++ src/main/java/io/emeraldpay/api/blockchain/Addresses.java
@@ -1,4 +1,4 @@
-package io.emeraldpay.grpc.blockchain;
+package io.emeraldpay.api.blockchain;
 
 import io.emeraldpay.api.proto.Common;
