+++ src/main/java/io/emeraldpay/api/blockchain/BlockchainRequest.java
@@ -1,4 +1,4 @@
-package io.emeraldpay.grpc.blockchain;
+package io.emeraldpay.api.blockchain;
 
 public class BlockchainRequest {
