+++ src/test/groovy/io/emeraldpay/api/BlockchainTypeSpec.groovy
@@ -1,4 +1,5 @@
-package io.emeraldpay.grpc
+package io.emeraldpay.api
+
 
 import spock.lang.Specification
