+++ src/test/groovy/io/emeraldpay/api/blockchain/AddressesSpec.groovy
@@ -1,4 +1,5 @@
-package io.emeraldpay.grpc.blockchain
+package io.emeraldpay.api.blockchain
+
 
 import io.emeraldpay.api.proto.Common
 import spock.lang.Specification
